refactor: move top level functions into aggregate.py
