+++ calculate/retirement.py
@@ -1,10 +1,8 @@
 from typing import Union
 from decimal import Decimal, ROUND_HALF_UP
-from calculate.federal_tax import (
-    calculate_annual_income_tax,
-    calculate_annual_federal_income_tax,
-    calculate_annual_state_income_tax,
-)
+from calculate.federal_tax import calculate_annual_federal_income_tax
+from calculate.state_tax import calculate_annual_state_income_tax
+from calculate.aggregate import calculate_annual_income_tax
 from calculate.state_capital_gains import get_state_capital_gains_calculator
 from utils.parameters import Person, Account, to_decimal
 from utils.enums import Frequency, MonthlyCompoundType, AccountType, State
