+++ main.py
@@ -4,7 +4,7 @@ import matplotlib.pyplot as plt
 import logging
 import json
 
-from calculate.federal_tax import (
+from calculate.aggregate import (
     calculate_income_distribution_data,
     calculate_retirement_deductions_excess,
 )
